+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/commands.py
@@ -14,12 +14,21 @@ from ._format import (
     webhook_get_config_output_format,
     webhook_list_events_output_format,
     webhook_ping_output_format,
-    replication_output_format
+    replication_output_format,
+    build_task_output_format,
+    build_task_detail_output_format,
+    build_output_format
 )
-from ._client_factory import cf_acr_registries, cf_acr_replications, cf_acr_webhooks
+from ._client_factory import (
+    cf_acr_registries,
+    cf_acr_replications,
+    cf_acr_webhooks,
+    cf_acr_build_tasks,
+    cf_acr_builds
+)
 
 
-def load_command_table(self, _):
+def load_command_table(self, _):  # pylint: disable=too-many-statements
 
     acr_custom_util = CliCommandType(
         operations_tmpl='azure.cli.command_modules.acr.custom#{}',
@@ -47,6 +56,18 @@ def load_command_table(self, _):
         operations_tmpl='azure.cli.command_modules.acr.replication#{}',
         table_transformer=replication_output_format,
         client_factory=cf_acr_replications
+    )
+
+    acr_build_util = CliCommandType(
+        operations_tmpl='azure.cli.command_modules.acr.build#{}',
+        table_transformer=build_output_format,
+        client_factory=cf_acr_builds
+    )
+
+    acr_build_task_util = CliCommandType(
+        operations_tmpl='azure.cli.command_modules.acr.build_task#{}',
+        table_transformer=build_task_detail_output_format,
+        client_factory=cf_acr_build_tasks
     )
 
     with self.command_group('acr', acr_custom_util) as g:
@@ -104,3 +125,21 @@ def load_command_table(self, _):
                                  custom_func_type=acr_replication_util,
                                  client_factory=cf_acr_replications,
                                  table_transformer=replication_output_format)
+
+    with self.command_group('acr', acr_build_util) as g:
+        g.command('build', 'acr_build')
+
+    with self.command_group('acr build-task', acr_build_task_util) as g:
+        g.command('create', 'acr_build_task_create')
+        g.command('show', 'acr_build_task_show')
+        g.command('list', 'acr_build_task_list', table_transformer=build_task_output_format)
+        g.command('delete', 'acr_build_task_delete')
+        g.command('update', 'acr_build_task_update')
+        g.command('run', 'acr_build_task_run', client_factory=cf_acr_builds,
+                  table_transformer=build_output_format)
+        g.command('list-builds', 'acr_build_task_list_builds', client_factory=cf_acr_builds,
+                  table_transformer=build_output_format)
+        g.command('show-build', 'acr_build_task_show_build', client_factory=cf_acr_builds,
+                  table_transformer=build_output_format)
+        g.command('logs', 'acr_build_task_logs', client_factory=cf_acr_builds,
+                  table_transformer=None)