+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -5,6 +5,8 @@ Release History
 
 2.0.24
 ++++++
+* Add ACR Build commands.
+* Improve resource not found error messages.
 * Improve resource creation performance and error handling.
 * Improve acr login in non-standard consoles and WSL.
 * Improve repository commands error messages.